requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
yt-dlp>=2024.12.13

openai>=1.51.0
pydantic>=2.7.0
aiofiles>=24.1.0
supabase>=2.0.0
pydub>=0.25.1
resend>=0.17.0


stripe>=8.0.0
coinbase-commerce>=1.0.1
python-dotenv>=1.0.0
httpx>=0.27.0
sentry-sdk[fastapi]
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
loguru>=0.7.2

# Backend specific dependencies (Refactor Update)
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-text-splitters>=0.0.1
langsmith>=0.1.0
stop-words==2018.7.23
langgraph-checkpoint-postgres>=2.0.0
testcontainers[postgres]>=3.7.1
langchain-community>=0.3.0
litellm>=1.0.0
pyyaml>=6.0.1
langfuse>=2.0.0
